refactored examples packages
