+++ android-terminal/src/test/java/electrostatic4j/aterminal/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.android_terminal;
+package electrostatic4j.aterminal;
 
 import org.junit.Test;
